Split domain aggregate and value object
